batata-project/requirements.txt
scikit-learn==0.22.1
pandas==1.0.1
minio==5.0.7
Django==2.0.2
djangorestframework==3.7.7
Pygments==2.6.1
